Import 2.6.31-3.2 release

+++ uclinux-rootfs/user/traceroute/linux-include/netinet/in_systm.h
@@ -47,9 +47,9 @@
  * represent the types with the bytes in ``high-ender'' order.
  */
 typedef u_short n_short;		/* short as received from the net */
-typedef u_long	n_long;			/* long as received from the net */
+typedef u_int	n_long;			/* long as received from the net */
 
-typedef	u_long	n_time;			/* ms since 00:00 GMT, byte rev */
+typedef	u_int	n_time;			/* ms since 00:00 GMT, byte rev */
 
 #ifdef KERNEL
 n_time	 iptime __P((void));
